+++ jolt-core/src/benches/bench.rs
@@ -1,7 +1,7 @@
 use crate::host;
 use crate::jolt::vm::rv32i_vm::{RV32IJoltVM, C, M};
 use crate::jolt::vm::Jolt;
-use crate::poly::structured_poly::HyraxConfig;
+use crate::poly::commitment::hyrax::HyraxScheme;
 use ark_bn254::G1Projective;
 use serde::Serialize;
 
@@ -67,11 +67,11 @@ fn prove_example<T: Serialize>(
 
         let preprocessing: crate::jolt::vm::JoltPreprocessing<
             ark_ff::Fp<ark_ff::MontBackend<ark_bn254::FrConfig, 4>, 4>,
-            HyraxConfig<ark_ec::short_weierstrass::Projective<ark_bn254::g1::Config>>,
+            HyraxScheme<ark_ec::short_weierstrass::Projective<ark_bn254::g1::Config>>,
         > = RV32IJoltVM::preprocess(bytecode.clone(), memory_init, 1 << 20, 1 << 20, 1 << 22);
 
         let (jolt_proof, jolt_commitments) =
-            <RV32IJoltVM as Jolt<_, HyraxConfig<G1Projective>, C, M>>::prove(
+            <RV32IJoltVM as Jolt<_, HyraxScheme<G1Projective>, C, M>>::prove(
                 io_device,
                 bytecode_trace,
                 memory_trace,
@@ -117,11 +117,11 @@ fn sha2chain() -> Vec<(tracing::Span, Box<dyn FnOnce()>)> {
 
         let preprocessing: crate::jolt::vm::JoltPreprocessing<
             ark_ff::Fp<ark_ff::MontBackend<ark_bn254::FrConfig, 4>, 4>,
-            HyraxConfig<ark_ec::short_weierstrass::Projective<ark_bn254::g1::Config>>,
+            HyraxScheme<ark_ec::short_weierstrass::Projective<ark_bn254::g1::Config>>,
         > = RV32IJoltVM::preprocess(bytecode.clone(), memory_init, 1 << 20, 1 << 20, 1 << 22);
 
         let (jolt_proof, jolt_commitments) =
-            <RV32IJoltVM as Jolt<_, HyraxConfig<G1Projective>, C, M>>::prove(
+            <RV32IJoltVM as Jolt<_, HyraxScheme<G1Projective>, C, M>>::prove(
                 io_device,
                 bytecode_trace,
                 memory_trace,
