+++ jolt-core/src/jolt/vm/rv32i_vm.rs
@@ -22,7 +22,7 @@ use crate::jolt::subtable::{
     srl::SrlSubtable, truncate_overflow::TruncateOverflowSubtable, xor::XorSubtable,
     JoltSubtableSet, LassoSubtable, SubtableId,
 };
-use crate::poly::structured_poly::CommitmentScheme;
+use crate::poly::commitment::commitment_scheme::CommitmentScheme;
 
 /// Generates an enum out of a list of JoltInstruction types. All JoltInstruction methods
 /// are callable on the enum type via enum_dispatch.
@@ -158,7 +158,7 @@ mod tests {
     use crate::host;
     use crate::jolt::instruction::JoltInstruction;
     use crate::jolt::vm::rv32i_vm::{Jolt, RV32IJoltVM, C, M};
-    use crate::poly::structured_poly::HyraxConfig;
+    use crate::poly::commitment::hyrax::HyraxScheme;
     use std::sync::Mutex;
     use strum::{EnumCount, IntoEnumIterator};
 
@@ -172,11 +172,11 @@ mod tests {
     fn instruction_set_subtables() {
         let mut subtable_set: HashSet<_> = HashSet::new();
         for instruction in
-            <RV32IJoltVM as Jolt<_, HyraxConfig<G1Projective>, C, M>>::InstructionSet::iter()
+            <RV32IJoltVM as Jolt<_, HyraxScheme<G1Projective>, C, M>>::InstructionSet::iter()
         {
             for (subtable, _) in instruction.subtables::<Fr>(C, M) {
                 // panics if subtable cannot be cast to enum variant
-                let _ = <RV32IJoltVM as Jolt<_, HyraxConfig<G1Projective>, C, M>>::Subtables::from(
+                let _ = <RV32IJoltVM as Jolt<_, HyraxScheme<G1Projective>, C, M>>::Subtables::from(
                     subtable.subtable_id(),
                 );
                 subtable_set.insert(subtable.subtable_id());
@@ -184,7 +184,7 @@ mod tests {
         }
         assert_eq!(
             subtable_set.len(),
-            <RV32IJoltVM as Jolt<_, HyraxConfig<G1Projective>, C, M>>::Subtables::COUNT,
+            <RV32IJoltVM as Jolt<_, HyraxScheme<G1Projective>, C, M>>::Subtables::COUNT,
             "Unused enum variants in Subtables"
         );
     }
@@ -202,7 +202,7 @@ mod tests {
         let preprocessing =
             RV32IJoltVM::preprocess(bytecode.clone(), memory_init, 1 << 20, 1 << 20, 1 << 20);
         let (proof, commitments) =
-            <RV32IJoltVM as Jolt<Fr, HyraxConfig<G1Projective>, C, M>>::prove(
+            <RV32IJoltVM as Jolt<Fr, HyraxScheme<G1Projective>, C, M>>::prove(
                 io_device,
                 bytecode_trace,
                 memory_trace,
@@ -231,7 +231,7 @@ mod tests {
         let preprocessing =
             RV32IJoltVM::preprocess(bytecode.clone(), memory_init, 1 << 20, 1 << 20, 1 << 20);
         let (jolt_proof, jolt_commitments) =
-            <RV32IJoltVM as Jolt<_, HyraxConfig<G1Projective>, C, M>>::prove(
+            <RV32IJoltVM as Jolt<_, HyraxScheme<G1Projective>, C, M>>::prove(
                 io_device,
                 bytecode_trace,
                 memory_trace,
