+++ jolt-core/src/poly/mod.rs
@@ -1,7 +1,7 @@
+pub mod commitment;
 pub mod dense_mlpoly;
 pub mod eq_poly;
-pub mod hyrax;
+pub mod field;
 pub mod identity_poly;
-pub mod pedersen;
 pub mod structured_poly;
 pub mod unipoly;
